app.js now allows for multiple employees to render

+++ app.js
@@ -15,6 +15,7 @@ const render = require("./lib/htmlRenderer");
 // Write code to use inquirer to gather information about the development team members,
 // and to create objects for each team member (using the correct classes as blueprints!)
 const employees =  []
+const createEmployees = () => {
 inquirer
   .prompt([
     {
@@ -52,9 +53,21 @@ inquirer
           ])
           .then(answer => {
             employees.push(new Engineer(answers.name, answers.id, answers.email, answer.github))
-            fs.writeFile(outputPath, render(employees), err => {
-              if (err) { console.log(err) }
-              console.log('File successfully created!')
+            inquirer.prompt({
+              type: 'confirm',
+              name: 'cont',
+              message: 'Continue?'
+            })
+            .then(({cont}) => {
+              if (cont) {
+                createEmployees()
+              }
+              else {
+                fs.writeFile(outputPath, render(employees), err => {
+                  if (err) { console.log(err) }
+                  console.log('File successfully created!')
+                })
+              }
             })
           })
           .catch (err => console.log(err))
@@ -70,10 +83,22 @@ inquirer
           ])
           .then(answer => {
             employees.push(new Intern(answers.name, answers.id, answers.email, answer.school))
-            fs.writeFile(outputPath, render(employees), err => {
-              if (err) { console.log(err) }
-              console.log('File successfully created!')
+            inquirer.prompt({
+              type: 'confirm',
+              name: 'cont',
+              message: 'Continue?'
             })
+              .then(({ cont }) => {
+                if (cont) {
+                  createEmployees()
+                }
+                else {
+                  fs.writeFile(outputPath, render(employees), err => {
+                    if (err) { console.log(err) }
+                    console.log('File successfully created!')
+                  })
+                }
+              })
           })
           .catch(err => console.log(err))
         break
@@ -88,16 +113,31 @@ inquirer
           ])
           .then(answer => {
             employees.push(new Manager(answers.name, answers.id, answers.email, answer.office))
-            fs.writeFile(outputPath, render(employees), err => {
-              if (err) { console.log(err) }
-              console.log('File successfully created!')
+            inquirer.prompt({
+              type: 'confirm',
+              name: 'cont',
+              message: 'Continue?'
             })
+              .then(({ cont }) => {
+                if (cont) {
+                  createEmployees()
+                }
+                else {
+                  fs.writeFile(outputPath, render(employees), err => {
+                    if (err) { console.log(err) }
+                    console.log('File successfully created!')
+                  })
+                }
+              })
           })
           .catch(err => console.log(err))
         break
     }
   })
   .catch(err => console.log(err))
+}
+
+createEmployees()
 
 // After the user has input all employees desired, call the `render` function (required
 // above) and pass in an array containing all employee objects; the `render` function will
